Separate `getProposalFields`, `getVoteFields`
