+++ packages/proposal-relayer/libs/utils/getProposalFields.ts
@@ -2,13 +2,10 @@ import type { EmbedFieldData } from "discord.js";
 import type {
 	ProposalDetails,
 	ProposalInfo,
-	ProposalVotes,
 	ProposalStatus,
-	VoteAction,
 } from "@proposal-relayer/libs/types";
 
-import { PROPOSALS_URL } from "@proposal-relayer/libs/constants";
-import { MessageEmbed, MessageActionRow, MessageButton } from "discord.js";
+import { MessageEmbed } from "discord.js";
 
 export function getProposalFields(
 	proposalDetails: ProposalDetails,
@@ -28,41 +25,6 @@ export function getProposalFields(
 			value: `${proposalInfo.enactmentDelay} blocks`,
 		},
 	];
-}
-
-export function getVoteFields({
-	passVotes,
-	rejectVotes,
-}: ProposalVotes): EmbedFieldData[] {
-	return [
-		{
-			name: "Votes to Pass",
-			value: `_**${passVotes}**_`,
-			inline: true,
-		},
-		{
-			name: "Votes to Reject",
-			value: `_**${rejectVotes}**_`,
-			inline: true,
-		},
-	];
-}
-
-export function getVoteButtons(proposalId: number): MessageActionRow {
-	return new MessageActionRow().addComponents(
-		new MessageButton()
-			.setURL(getProposalLink(proposalId, "pass"))
-			.setLabel("Pass")
-			.setStyle("LINK"),
-		new MessageButton()
-			.setURL(getProposalLink(proposalId, "reject"))
-			.setLabel("Reject")
-			.setStyle("LINK")
-	);
-}
-
-function getProposalLink(proposalId: number, action: VoteAction): string {
-	return `${PROPOSALS_URL}/${proposalId}?stage=proposal&action=${action}`;
 }
 
 export function getProposalEmbed(
